+++ src/GameEngine/Missile.java
@@ -19,7 +19,7 @@ public class Missile {
   /**
    * Define the current x position of the missile
    */
- private long id; //The missile id is it's creation's date
+ private int id; //The missile id is it's creation's date
  private Point position;
  private Unit target;
  private int damages;
@@ -31,7 +31,7 @@ public class Missile {
   /**
    * Constructor of the Missile class
    */
-  public Missile(long id, Tower origin, Point position, Unit target, double speed, int damages) {
+  public Missile(int id, Tower origin, Point position, Unit target, double speed, int damages) {
 	  System.out.println("Creation of a new missile !");
 	  this.id = id;
 	  this.position = position;
@@ -141,7 +141,7 @@ public class Missile {
    * Getter
    * @return
    */
-  public long getId(){
+  public int getId(){
 	  return this.id;
   }
   /**
